updating organization of api service
